Code cleaning and reformating
Some parts got optimized, some outputs got renewed

+++ src/main/java/com/ancientshores/Ancient/Classes/Spells/Commands/SkipCommand.java
@@ -11,7 +11,7 @@ public class SkipCommand extends ICommand {
     public boolean playCommand(EffectArgs ca) {
         try {
             if (ca.getParams().get(0) instanceof Number) {
-                int number = (int) ((Number) ca.getParams().get(0)).doubleValue();
+                int number = ((Number) ca.getParams().get(0)).intValue();
                 ca.getSpell().skipCommands(ca.getSpellInfo(), number);
                 return true;
             }
